do not use cache when explicitly loading the master list
We now have a BDS foreground service running all the time (if bluetooth detection is turned on), which prevents static members from being GCed and hence keeps LoaderCache in memory. This causes all the category tracks to always remains the same.
We explicitly release cache every time we are loading the master list