use id instead of trackPath/minor fixes and improvements
Track path is a content uri, with the difference being the id of the track. There's no need to store the entire path since it is same for all tracks, store only the unique bit which is the id.
Perform history deletion in order to prevent multiple threads accessing same files